Update: Restructured BatchMultiSigCall to contain classes
