+++ src/batchMultiSigCall/classes/EIP712/helpers.ts
@@ -1,0 +1,25 @@
+import { FCTCallParam, Param } from "types";
+
+export const getParams = (params: Param[]): Record<string, FCTCallParam> => {
+  return {
+    ...params.reduce((acc, param) => {
+      let value: FCTCallParam;
+
+      if (param.customType || param.type.includes("tuple")) {
+        if (param.type.lastIndexOf("[") > 0) {
+          const valueArray = param.value as Param[][];
+          value = valueArray.map((item) => getParams(item));
+        } else {
+          const valueArray = param.value as Param[];
+          value = getParams(valueArray);
+        }
+      } else {
+        value = param.value as string[] | string | boolean;
+      }
+      return {
+        ...acc,
+        [param.name]: value,
+      };
+    }, {}),
+  };
+};
