+++ src/batchMultiSigCall/classes/FCTBase.ts
@@ -1,0 +1,8 @@
+import { BatchMultiSigCall } from "methods";
+
+export class FCTBase {
+  public FCT: BatchMultiSigCall;
+  constructor(FCT: BatchMultiSigCall) {
+    this.FCT = FCT;
+  }
+}
