+++ src/batchMultiSigCall/helpers/handlers.ts
@@ -1,8 +1,8 @@
 import { utils } from "ethers";
 
 import { nullValue } from "../../constants";
-import { getEncodedMethodParams, getMethodInterface, getTypedHashes, getTypesArray } from "../../helpers";
-import { BatchMultiSigCallTypedData, IMSCallInput } from "../types";
+import { getEncodedMethodParams, getMethodInterface, getTypesArray } from "../../helpers";
+import { IMSCallInput } from "../types";
 
 export const handleMethodInterface = (call: IMSCallInput): string => {
   if (call.method) {
@@ -20,12 +20,12 @@ export const handleFunctionSignature = (call: IMSCallInput) => {
   return nullValue;
 };
 
-export const handleEnsHash = (call: IMSCallInput) => {
-  if (call.toENS) {
-    return utils.id(call.toENS);
-  }
-  return nullValue;
-};
+// export const handleEnsHash = (call: IMSCallInput) => {
+//   if (call.toENS) {
+//     return utils.id(call.toENS);
+//   }
+//   return nullValue;
+// };
 
 export const handleData = (call: IMSCallInput) => {
   return getEncodedMethodParams(call);
@@ -34,13 +34,6 @@ export const handleData = (call: IMSCallInput) => {
 export const handleTypes = (call: IMSCallInput) => {
   if (call.params) {
     return getTypesArray(call.params);
-  }
-  return [];
-};
-
-export const handleTypedHashes = (call: IMSCallInput, typedData: BatchMultiSigCallTypedData) => {
-  if (call.params) {
-    return getTypedHashes(call.params, typedData);
   }
   return [];
 };
