+++ src/batchMultiSigCall/index.ts
@@ -1,4 +1,3 @@
 export { BatchMultiSigCall } from "./batchMultiSigCall";
 export * from "./constants";
-export * as helpers from "./helpers";
 export * as utils from "./utils";
