+++ src/batchMultiSigCall/methods/index.ts
@@ -1,5 +1,3 @@
-export * from "./computed";
 export * from "./FCT";
 export * from "./helpers";
 export * from "./plugins";
-export * from "./variables";
